requirements.txt
Django==4.1.7
djangorestframework==3.14.0
django-filter==22.1
django-rest-passwordreset==1.3.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
requests==2.28.2
PyYAML==6.0
django-allauth==0.53.1
django-etc==1.4.0
redis==4.5.4
celery==5.2.7
drf-spectacular==0.26.1
drf-spectacular-sidecar==2023.3.1
gunicorn==20.1.0